services/voice-ai/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "caseconnect-voice-ai"
version = "0.1.0"
description = "Voice AI service for CaseConnect - voice-to-structured-data pipeline"
requires-python = ">=3.11"
license = {text = "Proprietary"}

dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "python-multipart>=0.0.9",
    "pydantic>=2.7.0",
    "numpy>=1.26.0",
    "openai-whisper>=20231117",
    "transformers>=4.40.0",
    "torch>=2.2.0",
    "spacy>=3.7.0",
    "websockets>=12.0",
    "python-dotenv>=1.0.1",
    "aiofiles>=23.2.1",
    "httpx>=0.27.0",
    "soundfile>=0.12.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
